Level 1 Created

+++ app/src/main/java/com/exe/paradox/Models/LevelData.java
@@ -3,7 +3,7 @@ package com.exe.paradox.Models;
 public class LevelData {
     public int levelNo;
     public boolean isLevelActive;
-    public long levelStartsInSeconds;
+    public double levelStartsInSeconds;
     public long levelStartsAt;
 
     public LevelData() {
